requirements.txt
gradio
selenium
webdriver-manager
beautifulsoup4
python-dotenv
google-generativeai